requirements.txt
SQLAlchemy
psycopg2-binary
python-dotenv
aiohttp
beautifulsoup4
feedparser
pytest
pytest-asyncio
alembic
pandas
pydantic
apache-airflow
apache-airflow-providers-postgres
sqlalchemy-utils
selenium
webdriver-manager
PyPDF2